requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect==1.2.0
cssutils==2.6.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-ninja==0.21.0
django-ratelimit==4.0.0
django-rest-auth==0.9.5
django-seed==0.3.1
django-storages==1.13.2
django-user-agents==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.5
Faker==17.6.0
frozenlist==1.3.3
geoip2==4.6.0
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
maxminddb==2.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
premailer==3.10.0
protobuf==4.23.1
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.6
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
ratelimit==2.2.1
regex==2022.10.31
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
toposort==1.10
typing_extensions==4.5.0
tzdata==2022.7
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
yagmail==0.15.293
yarl==1.9.2
